+++ HISTORY.rst
@@ -9,6 +9,7 @@ History
 * Added get_session_data()
 * Improved reliability of `FuncSeleniumMixin.get_literal_url()`
 * Allow <select> elements to be set using integers for values.
+* Fixed issues with `.value()` for radio buttons and text areas
 
 0.1.4
 +++++